Moved counter home page to app.dart.
